use a simpler and more sparse matrix for testing memory limits
Using a large, dense and completely random matrix with a sparse
solver takes a long time, but this test is only testing whether
MUMPSContext correctly increases the memory limit when it detects
that MUMPS has run out. This means we can choose a matrix that
is sparse and easy to factorize, as long as the memory requirements
are large.